refactor: modularize angular workspace
