+++ src/app/app.component.scss
@@ -1,0 +1,4 @@
+:host {
+  display: block;
+  min-height: 100vh;
+}
