--- src/app/app.scss
@@ -1,9 +1,0 @@
-.preview {
-  position: absolute;
-  transform: translate(-50%, -120%);
-  pointer-events: auto;
-}
-
-.preview span {
-  pointer-events: none;
-}
