+++ src/app/core/models/annotation.model.ts
@@ -1,0 +1,18 @@
+export type FieldType = 'text' | 'check' | 'radio' | 'number' | 'barcode' | 'signature';
+
+export interface PageField {
+  x: number;
+  y: number;
+  mapField: string;
+  fontSize: number;
+  color: string;
+  type: FieldType;
+  value?: string;
+  appender?: string;
+  decimals?: number | null;
+}
+
+export interface PageAnnotations {
+  num: number;
+  fields: PageField[];
+}
