+++ src/app/features/editor/components/sidebar/editor-sidebar.component.html
@@ -1,0 +1,117 @@
+<aside class="sidebar">
+  <section
+    class="sidebar-upload"
+    role="button"
+    tabindex="0"
+    (click)="requestPdfUpload.emit()"
+    (keydown)="uploadKeydown.emit($event)"
+    (dragover)="fileDragOver.emit($event)"
+    (dragleave)="fileDragLeave.emit($event)"
+    (drop)="fileDrop.emit($event)"
+    [class.sidebar-upload--active]="fileDropActive"
+    [attr.aria-label]="'app.workspaceUpload.ariaLabel' | t"
+  >
+    <input
+      #pdfFileInput
+      type="file"
+      accept="application/pdf"
+      (change)="pdfSelected.emit($event)"
+      hidden
+    />
+
+    <div class="sidebar-upload__body">
+      <div class="sidebar-upload__icon" aria-hidden="true">🗂️</div>
+      <div class="sidebar-upload__copy">
+        <span class="sidebar-upload__title">{{ 'app.workspaceUpload.title' | t }}</span>
+        <span class="sidebar-upload__subtitle">{{ 'app.workspaceUpload.subtitle' | t }}</span>
+      </div>
+    </div>
+    <span class="sidebar-upload__hint">{{ 'app.upload.hint' | t }}</span>
+  </section>
+
+  <h4>{{ 'sidebar.title' | t }}</h4>
+  <div class="sidebar-actions">
+    <button type="button" class="btn" (click)="triggerImportCoords.emit()">
+      {{ 'sidebar.importJsonFile' | t }}
+    </button>
+    <button type="button" class="btn" (click)="applyCoordsText.emit()">
+      {{ 'sidebar.applyJson' | t }}
+    </button>
+    <input
+      #coordsFileInput
+      type="file"
+      accept="application/json"
+      (change)="coordsFileSelected.emit($event)"
+      hidden
+    />
+  </div>
+
+  <section class="templates">
+    <h5>{{ 'sidebar.templates.title' | t }}</h5>
+    <div class="templates__save">
+      <input
+        type="text"
+        [ngModel]="templateName"
+        (ngModelChange)="onTemplateNameInput($event)"
+        [ngModelOptions]="{ standalone: true }"
+        [placeholder]="'sidebar.templates.placeholder' | t"
+      />
+      <button
+        type="button"
+        class="btn"
+        (click)="saveTemplate.emit()"
+        [disabled]="!hasAnnotations || !templateName.trim()"
+      >
+        {{ 'sidebar.templates.saveButton' | t }}
+      </button>
+    </div>
+    <ng-container *ngIf="templates.length; else emptyTemplates">
+      <div class="templates__select">
+        <label>{{ 'sidebar.templates.selectLabel' | t }}</label>
+        <div class="templates__controls">
+          <select
+            [ngModel]="selectedTemplateId"
+            (ngModelChange)="onSelectedTemplateChange($event)"
+            [ngModelOptions]="{ standalone: true }"
+          >
+            <option *ngFor="let template of templates" [value]="template.id">
+              {{ template.name }}
+            </option>
+          </select>
+          <button type="button" class="btn" (click)="loadTemplate.emit()" [disabled]="!selectedTemplateId">
+            {{ 'sidebar.templates.loadButton' | t }}
+          </button>
+          <button
+            type="button"
+            class="btn templates__delete"
+            (click)="deleteTemplate.emit()"
+            [disabled]="!selectedTemplateId || selectedTemplateId === defaultTemplateId"
+          >
+            {{ 'sidebar.templates.deleteButton' | t }}
+          </button>
+        </div>
+      </div>
+    </ng-container>
+    <ng-template #emptyTemplates>
+      <p class="templates__empty">{{ 'sidebar.templates.empty' | t }}</p>
+    </ng-template>
+  </section>
+
+  <textarea
+    class="json-editor"
+    [ngModel]="coordsText"
+    (ngModelChange)="coordsTextChange.emit($event)"
+    [ngModelOptions]="{ standalone: true }"
+    [placeholder]="'sidebar.jsonPlaceholder' | t"
+    spellcheck="false"
+  ></textarea>
+  <div class="json-toolbar">
+    <button class="btn" (click)="copyJson.emit()" [disabled]="!hasAnnotations">
+      {{ 'controls.copyJson' | t }}
+    </button>
+    <button class="btn" (click)="downloadJson.emit()" [disabled]="!hasAnnotations">
+      {{ 'controls.downloadJson' | t }}
+    </button>
+  </div>
+  <small class="sidebar-hint">{{ 'sidebar.jsonHint' | t }}</small>
+</aside>
