+++ src/app/features/editor/components/sidebar/editor-sidebar.component.scss
@@ -1,0 +1,135 @@
+:host {
+  display: block;
+  background: linear-gradient(180deg, rgba(15, 23, 42, 0.85), rgba(15, 23, 42, 0.65));
+  color: #f8fafc;
+  padding: 1.5rem;
+  width: 320px;
+  box-shadow: inset -1px 0 0 rgba(148, 163, 184, 0.08);
+}
+
+.sidebar {
+  display: flex;
+  flex-direction: column;
+  gap: 1.25rem;
+  height: 100%;
+}
+
+.sidebar-upload {
+  border: 1px dashed rgba(94, 234, 212, 0.35);
+  border-radius: 20px;
+  padding: 1.5rem;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  cursor: pointer;
+  transition: border-color 0.2s ease, background 0.2s ease, transform 0.2s ease;
+  background: rgba(15, 23, 42, 0.45);
+}
+
+.sidebar-upload:hover,
+.sidebar-upload--active {
+  border-color: rgba(94, 234, 212, 0.65);
+  background: rgba(15, 23, 42, 0.65);
+  transform: translateY(-1px);
+}
+
+.sidebar-upload__body {
+  display: flex;
+  gap: 0.75rem;
+  align-items: center;
+}
+
+.sidebar-upload__icon {
+  font-size: 2rem;
+}
+
+.sidebar-upload__copy {
+  display: flex;
+  flex-direction: column;
+  gap: 0.25rem;
+}
+
+.sidebar-upload__title {
+  font-weight: 600;
+}
+
+.sidebar-upload__subtitle,
+.sidebar-upload__hint {
+  font-size: 0.85rem;
+  color: rgba(226, 232, 240, 0.75);
+}
+
+.sidebar-actions {
+  display: flex;
+  gap: 0.75rem;
+}
+
+.templates {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  padding: 1rem;
+  border-radius: 16px;
+  background: rgba(15, 23, 42, 0.55);
+}
+
+.templates__save {
+  display: flex;
+  gap: 0.75rem;
+}
+
+.templates__save input,
+.templates__controls select,
+.json-editor {
+  width: 100%;
+  border-radius: 12px;
+  border: 1px solid rgba(148, 163, 184, 0.35);
+  background: rgba(15, 23, 42, 0.85);
+  color: inherit;
+  padding: 0.75rem 1rem;
+}
+
+.templates__controls {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.5rem;
+}
+
+.templates__delete {
+  background: rgba(248, 113, 113, 0.18);
+  border-color: rgba(248, 113, 113, 0.45);
+  color: #fecaca;
+}
+
+.templates__empty {
+  margin: 0;
+  font-size: 0.85rem;
+  color: rgba(226, 232, 240, 0.65);
+}
+
+.json-editor {
+  min-height: 220px;
+  font-family: 'Fira Code', monospace;
+  font-size: 0.85rem;
+  resize: vertical;
+}
+
+.json-toolbar {
+  display: flex;
+  gap: 0.75rem;
+}
+
+.sidebar-hint {
+  font-size: 0.75rem;
+  color: rgba(148, 163, 184, 0.65);
+}
+
+@media (max-width: 1080px) {
+  :host {
+    width: 100%;
+  }
+
+  .sidebar {
+    flex-direction: column;
+  }
+}
