+++ src/app/features/editor/components/toolbar/editor-toolbar.component.html
@@ -1,0 +1,69 @@
+<header class="toolbar" role="toolbar" [attr.aria-label]="'controls.toolbarAriaLabel' | t">
+  <div class="toolbar__brand">
+    <img class="toolbar__logo" src="/logo.svg" [attr.alt]="'Logotipo de ' + appName" />
+    <div class="toolbar__brand-text">
+      <span class="toolbar__app-name">{{ appName }}</span>
+      <span class="toolbar__version">v{{ version }}</span>
+    </div>
+  </div>
+
+  <div class="toolbar__group" role="group" [attr.aria-label]="'controls.navigationGroup' | t">
+    <button class="btn btn--icon" (click)="prevPage.emit()" [disabled]="!hasPdf || pageIndex === 1">
+      <span class="btn__icon" aria-hidden="true">⟨</span>
+      <span class="btn__label">{{ 'controls.prev' | t }}</span>
+    </button>
+    <span class="toolbar__badge">
+      {{ 'controls.pageIndicator' | t : { current: pageIndex, total: pageCount || 0 } }}
+    </span>
+    <button class="btn btn--icon" (click)="nextPage.emit()" [disabled]="!hasPdf || pageIndex === pageCount">
+      <span class="btn__label">{{ 'controls.next' | t }}</span>
+      <span class="btn__icon" aria-hidden="true">⟩</span>
+    </button>
+  </div>
+
+  <div class="toolbar__group" role="group" [attr.aria-label]="'controls.zoomGroup' | t">
+    <button class="btn btn--icon" (click)="zoomOut.emit()" [disabled]="!hasPdf">
+      <span class="btn__icon" aria-hidden="true">−</span>
+      <span class="btn__label">{{ 'controls.zoomOut' | t }}</span>
+    </button>
+    <span class="toolbar__badge">{{ 'controls.zoomLabel' | t }} {{ scale | number: '1.0-2' }}×</span>
+    <button class="btn btn--icon" (click)="zoomIn.emit()" [disabled]="!hasPdf">
+      <span class="btn__label">{{ 'controls.zoomIn' | t }}</span>
+      <span class="btn__icon" aria-hidden="true">＋</span>
+    </button>
+  </div>
+
+  <div class="toolbar__group" role="group" [attr.aria-label]="'controls.historyGroup' | t">
+    <button class="btn btn--icon" (click)="undo.emit()" [disabled]="!canUndo">
+      <span class="btn__icon" aria-hidden="true">↺</span>
+      <span class="btn__label">{{ 'controls.undo' | t }}</span>
+    </button>
+    <button class="btn btn--icon" (click)="redo.emit()" [disabled]="!canRedo">
+      <span class="btn__icon" aria-hidden="true">↻</span>
+      <span class="btn__label">{{ 'controls.redo' | t }}</span>
+    </button>
+    <button class="btn btn--icon btn--danger" (click)="clearAll.emit()" [disabled]="!hasAnnotations">
+      <span class="btn__icon" aria-hidden="true">🧹</span>
+      <span class="btn__label">{{ 'controls.clear' | t }}</span>
+    </button>
+  </div>
+
+  <div class="toolbar__group" role="group" [attr.aria-label]="'controls.exportGroup' | t">
+    <button
+      class="btn btn--icon"
+      (click)="downloadAnnotated.emit()"
+      [disabled]="!hasPdf || !hasAnnotations"
+    >
+      <span class="btn__icon" aria-hidden="true">⬇</span>
+      <span class="btn__label">{{ 'controls.downloadPdf' | t }}</span>
+    </button>
+  </div>
+
+  <app-language-selector
+    class="toolbar__language-selector"
+    [languages]="languages"
+    [selectedLanguage]="selectedLanguage"
+    (languageChange)="languageChange.emit($event)"
+    variant="compact"
+  ></app-language-selector>
+</header>
