+++ src/app/features/editor/components/toolbar/editor-toolbar.component.scss
@@ -1,0 +1,77 @@
+:host {
+  display: block;
+  background: linear-gradient(90deg, #0f172a, #1e293b);
+  color: #fff;
+  box-shadow: 0 2px 12px rgba(15, 23, 42, 0.35);
+}
+
+.toolbar {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+  padding: 1rem 2rem;
+  flex-wrap: wrap;
+}
+
+.toolbar__brand {
+  display: flex;
+  align-items: center;
+  gap: 0.75rem;
+  margin-right: 1rem;
+}
+
+.toolbar__logo {
+  width: 40px;
+  height: 40px;
+}
+
+.toolbar__brand-text {
+  display: flex;
+  flex-direction: column;
+  line-height: 1;
+}
+
+.toolbar__app-name {
+  font-weight: 600;
+}
+
+.toolbar__version {
+  font-size: 0.75rem;
+  opacity: 0.7;
+}
+
+.toolbar__group {
+  display: inline-flex;
+  align-items: center;
+  gap: 0.5rem;
+  padding: 0.5rem 0.75rem;
+  background: rgba(255, 255, 255, 0.08);
+  border-radius: 999px;
+}
+
+.toolbar__badge {
+  padding: 0.25rem 0.75rem;
+  border-radius: 999px;
+  background: rgba(15, 23, 42, 0.4);
+  font-weight: 500;
+}
+
+.toolbar__language-selector {
+  margin-left: auto;
+}
+
+@media (max-width: 960px) {
+  .toolbar {
+    flex-direction: column;
+    align-items: stretch;
+  }
+
+  .toolbar__brand {
+    justify-content: center;
+  }
+
+  .toolbar__language-selector {
+    margin-left: 0;
+    align-self: center;
+  }
+}
