+++ src/app/features/editor/pages/editor-home/editor-home.page.scss
@@ -1,0 +1,163 @@
+:host {
+  display: block;
+}
+
+.app {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+  background: radial-gradient(circle at top, rgba(15, 23, 42, 0.45), transparent 60%),
+    #0f172a;
+  color: #e2e8f0;
+}
+
+.workspace {
+  display: grid;
+  grid-template-columns: 320px 1fr;
+  gap: 1.5rem;
+  padding: 1.5rem;
+  flex: 1 1 auto;
+}
+
+.workspace__sidebar {
+  grid-row: 1;
+}
+
+.viewer {
+  position: relative;
+  background: rgba(15, 23, 42, 0.55);
+  border-radius: 24px;
+  padding: 1.5rem;
+  overflow: hidden;
+  box-shadow: 0 18px 48px rgba(15, 23, 42, 0.45);
+}
+
+.canvas-container {
+  position: relative;
+  width: 100%;
+  max-height: calc(100vh - 220px);
+  overflow: auto;
+  border-radius: 18px;
+  background: #111827;
+}
+
+.canvas-container canvas {
+  display: block;
+  margin: 0 auto;
+}
+
+.canvas-container canvas + canvas,
+.canvas-container .annotations-layer {
+  position: absolute;
+  top: 0;
+  left: 50%;
+  transform: translateX(-50%);
+}
+
+.annotations-layer {
+  pointer-events: none;
+}
+
+.hitbox {
+  position: absolute;
+  inset: 0;
+  cursor: crosshair;
+}
+
+.annotation-editor {
+  position: absolute;
+  transform: translate(-50%, -100%);
+  background: rgba(15, 23, 42, 0.92);
+  padding: 1rem;
+  border-radius: 16px;
+  box-shadow: 0 20px 40px rgba(15, 23, 42, 0.45);
+  width: 260px;
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+}
+
+.annotation-editor--editing {
+  border: 1px solid rgba(59, 130, 246, 0.4);
+}
+
+.field {
+  display: flex;
+  flex-direction: column;
+  gap: 0.35rem;
+}
+
+.field-label {
+  font-size: 0.75rem;
+  font-weight: 600;
+  color: rgba(226, 232, 240, 0.8);
+}
+
+.field input,
+.field select {
+  padding: 0.45rem 0.6rem;
+  border-radius: 10px;
+  border: 1px solid rgba(148, 163, 184, 0.35);
+  background: rgba(15, 23, 42, 0.75);
+  color: inherit;
+}
+
+.editor-actions {
+  display: flex;
+  justify-content: space-between;
+  gap: 0.5rem;
+}
+
+.empty-state {
+  flex: 1;
+  display: grid;
+  place-items: center;
+  padding: 4rem 1.5rem;
+}
+
+.empty-state__content {
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  max-width: 420px;
+  background: rgba(15, 23, 42, 0.6);
+  padding: 2rem;
+  border-radius: 24px;
+  box-shadow: 0 18px 48px rgba(15, 23, 42, 0.45);
+}
+
+.empty-state__logo {
+  width: 72px;
+  height: 72px;
+  align-self: center;
+}
+
+.empty-state__title {
+  margin: 0;
+}
+
+.empty-state__description {
+  margin: 0;
+  color: rgba(226, 232, 240, 0.75);
+}
+
+@media (max-width: 1200px) {
+  .workspace {
+    grid-template-columns: 1fr;
+  }
+
+  .workspace__sidebar {
+    order: 2;
+  }
+}
+
+@media (max-width: 768px) {
+  .annotation-editor {
+    position: fixed;
+    left: 50%;
+    top: unset;
+    bottom: 2rem;
+    transform: translateX(-50%);
+  }
+}
