+++ src/app/features/landing/landing-routing.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { LandingHomePageComponent } from './pages/home/home.page';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: LandingHomePageComponent,
+  },
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class LandingRoutingModule {}
