+++ src/app/features/landing/pages/home/home.page.html
@@ -1,0 +1,51 @@
+<section class="landing">
+  <div class="landing__language">
+    <app-language-selector
+      [languages]="languages"
+      [selectedLanguage]="languageModel"
+      (languageChange)="onLanguageChange($event)"
+      variant="compact"
+    ></app-language-selector>
+  </div>
+
+  <div class="landing__hero">
+    <span class="landing__badge">{{ 'app.landing.badge' | t }}</span>
+    <h1 class="landing__title">{{ 'app.landing.title' | t }}</h1>
+    <p class="landing__description">{{ 'app.landing.description' | t }}</p>
+  </div>
+
+  <div
+    class="landing__dropzone"
+    role="button"
+    tabindex="0"
+    (click)="navigateToEditor()"
+    (keydown)="onDropZoneKeydown($event)"
+    (dragover)="onFileDragOver($event)"
+    (dragleave)="onFileDragLeave($event)"
+    (drop)="onFileDrop($event)"
+    [class.landing__dropzone--active]="fileDropActive()"
+    [attr.aria-label]="'app.upload.title' | t"
+  >
+    <div class="landing__icon" aria-hidden="true">📄</div>
+    <div class="landing__text">
+      <span class="landing__headline">{{ 'app.upload.title' | t }}</span>
+      <span class="landing__subheadline">{{ 'app.upload.subtitle' | t }}</span>
+    </div>
+    <button type="button" class="landing__cta">{{ 'app.landing.cta' | t }}</button>
+    <span class="landing__hint">{{ 'app.upload.hint' | t }}</span>
+  </div>
+
+  <footer class="landing__footer">
+    <div class="footer__brand">
+      <img class="footer__logo" src="/logo.svg" [attr.alt]="'Logotipo de ' + appName" />
+      <div class="footer__brand-text">
+        <span class="footer__app-name">{{ appName }}</span>
+        <span class="footer__version">v{{ version }}</span>
+      </div>
+    </div>
+    <div class="footer__details">
+      <span>© {{ currentYear }} {{ appName }}</span>
+      <span>Desarrollado por {{ appAuthor }}</span>
+    </div>
+  </footer>
+</section>
