+++ src/app/features/landing/pages/home/home.page.scss
@@ -1,0 +1,186 @@
+:host {
+  display: flex;
+  justify-content: center;
+  padding: 4rem 1.5rem;
+  min-height: 100vh;
+  background: radial-gradient(circle at top, rgba(94, 234, 212, 0.12), transparent 60%), #0f172a;
+  color: #f8fafc;
+}
+
+.landing {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+  gap: 2.5rem;
+  max-width: 720px;
+  width: 100%;
+  position: relative;
+}
+
+.landing__language {
+  position: absolute;
+  top: 0;
+  right: 0;
+  display: flex;
+  justify-content: flex-end;
+}
+
+.landing__hero {
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+}
+
+.landing__badge {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  padding: 0.4rem 1.2rem;
+  border-radius: 999px;
+  background: rgba(94, 234, 212, 0.15);
+  color: #5eead4;
+  font-weight: 600;
+  letter-spacing: 0.08em;
+  text-transform: uppercase;
+  font-size: 0.75rem;
+}
+
+.landing__title {
+  margin: 0;
+  font-size: clamp(2.5rem, 6vw, 3.5rem);
+}
+
+.landing__description {
+  margin: 0;
+  color: rgba(203, 213, 225, 0.85);
+  line-height: 1.6;
+}
+
+.landing__dropzone {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 1.25rem;
+  width: min(520px, 100%);
+  padding: 2.5rem;
+  border-radius: 28px;
+  border: 1px dashed rgba(94, 234, 212, 0.3);
+  background: rgba(15, 23, 42, 0.75);
+  box-shadow: 0 24px 48px rgba(5, 5, 17, 0.45);
+  cursor: pointer;
+  transition: border-color 0.2s ease, transform 0.2s ease, background 0.2s ease;
+  outline: none;
+}
+
+.landing__dropzone:hover {
+  transform: translateY(-2px);
+  border-color: rgba(94, 234, 212, 0.6);
+  background: rgba(32, 42, 62, 0.85);
+}
+
+.landing__dropzone:focus-visible {
+  box-shadow: 0 0 0 3px rgba(94, 234, 212, 0.35);
+}
+
+.landing__dropzone--active {
+  border-color: #5eead4;
+  background: rgba(15, 35, 40, 0.85);
+}
+
+.landing__icon {
+  display: grid;
+  place-items: center;
+  width: 84px;
+  height: 84px;
+  border-radius: 24px;
+  background: rgba(94, 234, 212, 0.12);
+  font-size: 2.8rem;
+}
+
+.landing__text {
+  display: flex;
+  flex-direction: column;
+  gap: 0.4rem;
+}
+
+.landing__headline {
+  font-size: 1.35rem;
+  font-weight: 600;
+}
+
+.landing__subheadline,
+.landing__hint {
+  color: rgba(203, 213, 225, 0.75);
+  font-size: 0.95rem;
+}
+
+.landing__cta {
+  padding: 0.6rem 1.6rem;
+  border-radius: 999px;
+  border: none;
+  font-weight: 600;
+  background: #5eead4;
+  color: #042424;
+  cursor: pointer;
+  transition: transform 0.2s ease, box-shadow 0.2s ease;
+}
+
+.landing__cta:hover {
+  transform: translateY(-1px);
+  box-shadow: 0 14px 26px rgba(94, 234, 212, 0.35);
+}
+
+.landing__footer {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+  align-items: center;
+  color: rgba(203, 213, 225, 0.75);
+}
+
+.footer__brand {
+  display: inline-flex;
+  align-items: center;
+  gap: 0.75rem;
+}
+
+.footer__logo {
+  width: 40px;
+  height: 40px;
+}
+
+.footer__brand-text {
+  display: flex;
+  flex-direction: column;
+  line-height: 1;
+}
+
+.footer__app-name {
+  font-weight: 600;
+}
+
+.footer__version {
+  font-size: 0.8rem;
+  opacity: 0.75;
+}
+
+.footer__details {
+  display: flex;
+  gap: 1rem;
+  flex-wrap: wrap;
+  justify-content: center;
+}
+
+@media (max-width: 720px) {
+  .landing__language {
+    position: static;
+    align-self: flex-end;
+  }
+
+  .landing__dropzone {
+    padding: 2rem 1.5rem;
+  }
+}
