--- src/app/i18n/translations/es-ES.json
@@ -1,114 +1,0 @@
-{
-  "app": {
-    "title": "PDF Annotator",
-    "languageLabel": "Idioma",
-    "languageAriaLabel": "Selecciona el idioma de la aplicación",
-    "languages": {
-      "es-ES": "Español",
-      "en": "Inglés",
-      "ca": "Catalán"
-    },
-    "landing": {
-      "badge": "Empieza a anotar",
-      "title": "Maqueta coordenadas de cualquier PDF en segundos",
-      "description": "Sube tu documento para añadir anotaciones vinculadas a JSON, reutilizar plantillas y exportar archivos listos para trabajar.",
-      "cta": "Seleccionar PDF"
-    },
-    "upload": {
-      "title": "Carga tu PDF",
-      "subtitle": "Arrastra y suelta el archivo aquí o haz clic para buscar",
-      "hint": "Solo PDF",
-      "invalidFormat": "Solo se admiten archivos PDF."
-    },
-    "workspaceUpload": {
-      "title": "¿Quieres empezar de cero?",
-      "subtitle": "Arrastra otro PDF o haz clic para sustituir el documento actual.",
-      "ariaLabel": "Subir un nuevo PDF para reiniciar el espacio de trabajo"
-    }
-  },
-  "controls": {
-    "prev": "Anterior",
-    "next": "Siguiente",
-    "pageIndicator": "Página {{current}} / {{total}}",
-    "zoomLabel": "Zoom",
-    "zoomIn": "Acercar",
-    "zoomOut": "Alejar",
-    "undo": "Deshacer",
-    "redo": "Rehacer",
-    "clear": "Limpiar",
-    "copyJson": "Copiar JSON",
-    "importJson": "Importar JSON",
-    "downloadJson": "Descargar JSON",
-    "downloadPdf": "Descargar PDF",
-    "toolbarAriaLabel": "Acciones del PDF",
-    "navigationGroup": "Navegación de páginas",
-    "zoomGroup": "Controles de zoom",
-    "historyGroup": "Acciones de historial",
-    "exportGroup": "Opciones de exportación"
-  },
-  "sidebar": {
-    "title": "Anotaciones (JSON)",
-    "importJsonFile": "Importar desde archivo",
-    "applyJson": "Aplicar JSON",
-    "jsonPlaceholder": "Pega o edita aquí las coordenadas de anotaciones...",
-    "templates": {
-      "title": "Plantillas reutilizables",
-      "placeholder": "Nombre de la plantilla",
-      "saveButton": "Guardar plantilla",
-      "selectLabel": "Plantillas guardadas",
-      "loadButton": "Cargar",
-      "deleteButton": "Eliminar",
-      "empty": "Aún no tienes plantillas guardadas."
-    },
-    "jsonHint": "Pega las coordenadas o impórtalas desde un JSON con el mismo formato que el archivo exportado."
-  },
-  "annotation": {
-    "fields": {
-      "text": {
-        "label": "Texto",
-        "placeholder": "Texto..."
-      },
-      "size": {
-        "label": "Tamaño"
-      },
-      "color": {
-        "label": "Color"
-      },
-      "type": {
-        "label": "Tipo",
-        "options": {
-          "text": "Texto",
-          "check": "Casilla",
-          "radio": "Botón de opción",
-          "number": "Número"
-        }
-      },
-      "value": {
-        "label": "Valor fijo",
-        "placeholder": "Valor opcional..."
-      },
-      "number": {
-        "decimals": {
-          "label": "Decimales"
-        },
-        "appender": {
-          "label": "Sufijo",
-          "placeholder": "Ej.: %"
-        }
-      }
-    },
-    "color": {
-      "hintPreview": "Selecciona un color, escribe un HEX o un RGB. El JSON siempre exportará el color en HEX.",
-      "hintEdit": "Puedes pegar un HEX o un RGB existente para reutilizar colores exactos.",
-      "hexPlaceholder": "#5eead4",
-      "rgbPlaceholder": "rgb(94, 234, 212)"
-    },
-    "actions": {
-      "duplicate": "Duplicar anotación"
-    },
-    "editingBadge": "Editando"
-  },
-  "viewer": {
-    "emptyMessage": "Sube un PDF y haz click donde quieras añadir anotaciones. Edítalas en la vista previa antes de confirmar ✅."
-  }
-}
