+++ src/app/shared/pipes/translation.pipe.ts
@@ -1,0 +1,15 @@
+import { Pipe, PipeTransform } from '@angular/core';
+
+import { TranslationService } from '../../core/services/translation.service';
+
+@Pipe({
+  name: 't',
+  pure: false,
+})
+export class TranslationPipe implements PipeTransform {
+  constructor(private readonly translation: TranslationService) {}
+
+  transform(key: string, params?: Record<string, unknown>): string {
+    return this.translation.translate(key, params);
+  }
+}
